[FileRepo] * Added getReadOnlyReason() function to FileRepo to check read-only status. Added such checks to the File object functions that mutate files in storage. This should make read-only mode more tolerable (which is needed at least briefly when switching backends). * Added lock()/unlock() calls to File restore() function. * Use proper isOK() accessor for Status objects.

+++ includes/filerepo/file/LocalFile.php
@@ -908,9 +908,13 @@ class LocalFile extends File {
 	 */
 	function upload( $srcPath, $comment, $pageText, $flags = 0, $props = false, $timestamp = false, $user = null ) {
 		global $wgContLang;
+
+		if ( $this->getRepo()->getReadOnlyReason() !== false ) {
+			return $this->readOnlyFatalStatus();
+		}
+
 		// truncate nicely or the DB will do it for us
-		// non-nicely (dangling multi-byte chars, non-truncated
-		// version in cache).
+		// non-nicely (dangling multi-byte chars, non-truncated version in cache).
 		$comment = $wgContLang->truncate( $comment, 255 );
 		$this->lock(); // begin
 		$status = $this->publish( $srcPath, $flags );
@@ -1175,6 +1179,10 @@ class LocalFile extends File {
 	 *     archive name, or an empty string if it was a new file.
 	 */
 	function publishTo( $srcPath, $dstRel, $flags = 0 ) {
+		if ( $this->getRepo()->getReadOnlyReason() !== false ) {
+			return $this->readOnlyFatalStatus();
+		}
+
 		$this->lock(); // begin
 
 		$archiveName = wfTimestamp( TS_MW ) . '!'. $this->getName();
@@ -1211,6 +1219,10 @@ class LocalFile extends File {
 	 * @return FileRepoStatus object.
 	 */
 	function move( $target ) {
+		if ( $this->getRepo()->getReadOnlyReason() !== false ) {
+			return $this->readOnlyFatalStatus();
+		}
+
 		wfDebugLog( 'imagemove', "Got request to move {$this->name} to " . $target->getText() );
 		$this->lock(); // begin
 
@@ -1250,6 +1262,10 @@ class LocalFile extends File {
 	 * @return FileRepoStatus object.
 	 */
 	function delete( $reason, $suppress = false ) {
+		if ( $this->getRepo()->getReadOnlyReason() !== false ) {
+			return $this->readOnlyFatalStatus();
+		}
+
 		$this->lock(); // begin
 
 		$batch = new LocalFileDeleteBatch( $this, $reason, $suppress );
@@ -1266,7 +1282,7 @@ class LocalFile extends File {
 		}
 		$status = $batch->execute();
 
-		if ( $status->ok ) {
+		if ( $status->isOK() ) {
 			// Update site_stats
 			$site_stats = $dbw->tableName( 'site_stats' );
 			$dbw->query( "UPDATE $site_stats SET ss_images=ss_images-1", __METHOD__ );
@@ -1293,6 +1309,10 @@ class LocalFile extends File {
 	 * @return FileRepoStatus object.
 	 */
 	function deleteOld( $archiveName, $reason, $suppress = false ) {
+		if ( $this->getRepo()->getReadOnlyReason() !== false ) {
+			return $this->readOnlyFatalStatus();
+		}
+
 		$this->lock(); // begin
 
 		$batch = new LocalFileDeleteBatch( $this, $reason, $suppress );
@@ -1302,7 +1322,7 @@ class LocalFile extends File {
 
 		$this->unlock(); // done
 
-		if ( $status->ok ) {
+		if ( $status->isOK() ) {
 			$this->purgeDescription();
 			$this->purgeHistory();
 		}
@@ -1322,6 +1342,12 @@ class LocalFile extends File {
 	 * @return FileRepoStatus
 	 */
 	function restore( $versions = array(), $unsuppress = false ) {
+		if ( $this->getRepo()->getReadOnlyReason() !== false ) {
+			return $this->readOnlyFatalStatus();
+		}
+
+		$this->lock(); // begin
+
 		$batch = new LocalFileRestoreBatch( $this, $unsuppress );
 
 		if ( !$versions ) {
@@ -1332,14 +1358,14 @@ class LocalFile extends File {
 
 		$status = $batch->execute();
 
-		if ( !$status->isGood() ) {
-			return $status;
+		if ( $status->isGood() ) {
+			$cleanupStatus = $batch->cleanup();
+			$cleanupStatus->successCount = 0;
+			$cleanupStatus->failCount = 0;
+			$status->merge( $cleanupStatus );
 		}
 
-		$cleanupStatus = $batch->cleanup();
-		$cleanupStatus->successCount = 0;
-		$cleanupStatus->failCount = 0;
-		$status->merge( $cleanupStatus );
+		$this->unlock(); // done
 
 		return $status;
 	}
@@ -1443,6 +1469,14 @@ class LocalFile extends File {
 		$this->locked = false;
 		$dbw = $this->repo->getMasterDB();
 		$dbw->rollback( __METHOD__ );
+	}
+
+	/**
+	 * @return Status 
+	 */
+	protected function readOnlyFatalStatus() {
+		return $this->getRepo()->newFatal( 'filereadonlyerror', $this->getName(),
+			$this->getRepo()->getName(), $this->getRepo()->getReadOnlyReason() );
 	}
 } // LocalFile class
 
@@ -1711,7 +1745,7 @@ class LocalFileDeleteBatch {
 			$this->status->merge( $status );
 		}
 
-		if ( !$this->status->ok ) {
+		if ( !$this->status->isOK() ) {
 			// Critical file deletion error
 			// Roll back inserts, release lock and abort
 			// TODO: delete the defunct filearchive rows if we are using a non-transactional DB
